More rigorous test framework and fix tests
The unit tests didn't rigorously test expectations. This led to tests
which were broken but passed anyway. For example for the git Mocks you
could add expectations for X, Y, and Z and if only X and Y were called
the tests would still pass. For the github mocks the same issue existed
but also they would pass if X and Y expectations were set but X and Y
were called multiple times. This change fixes both of these issues and
fixes the unit tests so they are no longer broken.

commit-id:495c3a84