Enable stricter expectations for github tests.
The application code executes some github commands in parallel. This
makes the order of some github non-deterministic. The existing
implementation allows for all commands to be executed in any order.
Test would be clearer with deterministic expectations. This runs each
test twice. One time they are executed without goroutines so the
expectations are in deterministic order. The second time they are
executed with goroutines to test the code's as it will be executed by
the application.

commit-id:fc104ef7