Start support for PR sets.
PR sets will allow you to manage several sets of PRs on a single branch.
This avoids branch switching and simply groups related PRs on the same
branch. You can of course continue to work on separate branches if
desired.

This features is still a WIP.

Included in this commit:
  * Capture state of commits (not just pull requests).
  * include commit index in spr status. This will be used by update to
    select commits to be a part of a commit set.
  * Update commit messages to have commit-id's

commit-id:370260d8