PR sets: track PR sets and include in spr status
PR sets will allow you to manage several sets of PRs on a single branch.
This avoids branch switching and simply groups related PRs on the same
branch. You can of course continue to work on separate branches if
desired.

This features is still a WIP.

Included in this commit:
  * Track which pull requests are a part of a pull request set (in .spr.state).
  * Include the pull request set in the spr status output.

commit-id:0b2762ab