PR sets: tracked orphaned PR sets
PR sets will allow you to manage several sets of PRs on a single branch.
This avoids branch switching and simply groups related PRs on the same
branch. You can of course continue to work on separate branches if
desired.

This features is still a WIP.

Included in this commit:
  * Track which pull requests are no longer referenced by a local commit and are therefore eligible for deletion.